pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "forge-core"
version = "0.1.0"
description = "A review-gated context compiler. Turn long-term content into multi-runtime agent context, with PR/approve gates and before/after evaluation."
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [{ name = "dxy", email = "peter.dong@gmail.com" }]
keywords = ["agent", "context", "compiler", "llm", "claude", "review-gated"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "pyyaml>=6.0",
    "click>=8.1",
]

[project.optional-dependencies]
dev = ["pytest>=7.0", "pytest-cov"]

[project.scripts]
forge = "forge.cli:main"

[tool.setuptools.packages.find]
include = ["forge*"]
exclude = ["tests*", "examples*"]
